Mocking with ts-mockito
Unit tests should test a unit in isolation. If you need to set up a bunch of other stuff then it's probably turning into an integration test.

Mocks can be used to ensure we're testing the unit in isolation. Plus, they eliminate most of the effort of having to set up dependencies.